requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
pynacl==1.5.0
websockets==11.0.3
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.1
